requirements.txt
numpy==1.16.1
scipy==1.4.1
pandas==1.0.4
xarray==0.15.1
numpy-quaternion==2020.9.5.14.42.2
numba==0.48.0
llvmlite==0.31.0
anytree==2.8.0
pooch==1.3.0
